BREAK (Commit messages) Remove colon from subject line
Experience shows that sometimes we forget the colon after the
(Component) element of the commit message subject line. So in the
tradition of paving the cowpaths, we remove it from the standard. It
also saves us one extra character, and does not impact readability (some
might argue that it improves it).