FEAT Add @see, @author and @version
From work on the haproxy-redirects file it is clear that it would be helpful to record additional information on the code blocks.

This additional allows us to see references (e.g. UniDesk calls), who added the code, and an internal version number for the code itself.

Restrict TOC depth.
Bump version number.